chore: refactor code (see description)
- removed V3 from wrap and wraps
- shifted sign from shared to v2 since it's only used in v2
- added index.ts to sign
- added @context attribute to v3's e2e tests
