+++ src/v2/sign/ecdsa-secp256k1-signature-2019.ts
@@ -4,7 +4,7 @@ import {
   SignedWrappedDocument,
   ProofPurpose,
   ProofType
-} from "../@types/document";
+} from "../../shared/@types/document";
 import { ethers } from "ethers";
 const type = ProofType.EcdsaSecp256k1Signature2019;
